tabs.tsx -> tab.tsx, tab.tsx, tab.tsx, tab.tsx
