+++ src/components/pages/movie-page.tsx
@@ -5,7 +5,10 @@ import {Overviews} from '../../mocks/overview';
 import {AppRoute, getReviewRoute} from '../app';
 import {Cards, getSimilarMovies } from '../film-card';
 import {Link, useNavigate, useParams} from 'react-router-dom';
-import { DetailsTab, OverviewTab, ReviewsTab, TabsNavigation } from '../tabs';
+import { TabsNavigation } from '../tab-navigation';
+import { DetailsTab } from '../tab-details';
+import { OverviewTab } from '../tab-overviews';
+import { ReviewsTab } from '../tab-reviews';
 import { Reviews } from '../../mocks/reviews';
 import { Footer } from '../footer';
